agent/pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "vigilops-agent"
version = "0.1.0"
description = "VigilOps monitoring agent"
requires-python = ">=3.9"
dependencies = [
    "psutil>=5.9",
    "httpx>=0.27",
    "pyyaml>=6.0",
    "click>=8.1",
]

[project.optional-dependencies]
postgres = ["psycopg2-binary>=2.9"]
mysql = ["pymysql>=1.1"]
oracle = ["oracledb>=2.0"]
redis = ["redis>=4.0"]
db = [
    "psycopg2-binary>=2.9",
    "pymysql>=1.1",
    "oracledb>=2.0",
    "redis>=4.0",
]

[project.scripts]
vigilops-agent = "vigilops_agent.cli:main"

[dependency-groups]
dev = [
    "pytest>=8.4.2",
]
